refactor: 코드 수정

+++ src/main/java/gift/repository/ProductRepository.java
@@ -121,7 +121,7 @@ public class ProductRepository {
             new Product(
                 rs.getLong("id"),
                 rs.getString("name"),
-                rs.getInt("price"),
+                rs.getLong("price"),
                 rs.getString("image_url")
             )
         );
